Refactor ImportService to use strategy pattern for scanners
- Extracted platform-specific scanning logic into dedicated scanner classes in `main/scanners/`.
- Created `GameScanner` interface and `BaseGameScanner` abstract class.
- Implemented scanners for Steam, Xbox, Epic, GOG, Ubisoft, Rockstar, EA, Battle.net, Humble, Itch, Manual folders, and Generic Deep Scan.
- Refactored `ImportService.ts` to orchestrate these scanners, significantly reducing complexity (from ~1900 lines to ~280 lines).
- Moved `ScannedGameResult` to `main/scanners/ScannedGameResult.ts` to avoid circular dependencies, while re-exporting it from `ImportService.ts` for backward compatibility.
- Fixed an existing type error in `main/GameMatcher.test.ts` discovered during build verification.
