+++ main/ImportService.ts
@@ -1,30 +1,27 @@
-import { SteamService, SteamGame } from './SteamService.js';
-import { XboxService, XboxGame } from './XboxService.js';
+import { SteamService } from './SteamService.js';
+import { XboxService } from './XboxService.js';
 import { AppConfigService } from './AppConfigService.js';
 import { MetadataFetcherService } from './MetadataFetcherService.js';
 import { GameFilteringService } from './GameFilteringService.js';
-import { existsSync, readdirSync, readFileSync, statSync } from 'node:fs';
-import { join, sep, dirname } from 'node:path';
-import { execSync } from 'node:child_process';
+import { existsSync } from 'node:fs';
 
-export interface ScannedGameResult {
-  uuid: string;
-  source: string;
-  originalName: string;
-  installPath: string;
-  exePath?: string;
-  launchArgs?: string;  // Command-line arguments for exe launch
-  appId?: string;
-  packageFamilyName?: string;
-  appUserModelId?: string;
-  launchUri?: string;
-  xboxKind?: 'uwp' | 'pc';
-  title: string;
-  status: 'pending' | 'scanning' | 'matched' | 'ambiguous' | 'ready' | 'error';
-  categories?: string[];
-  isDownloading?: boolean;
-  error?: string;
-}
+// Import Scanners
+import { ScannedGameResult } from './scanners/ScannedGameResult.js';
+import { SteamScanner } from './scanners/SteamScanner.js';
+import { XboxScanner } from './scanners/XboxScanner.js';
+import { EpicScanner } from './scanners/EpicScanner.js';
+import { GogScanner } from './scanners/GogScanner.js';
+import { UbisoftScanner } from './scanners/UbisoftScanner.js';
+import { RockstarScanner } from './scanners/RockstarScanner.js';
+import { EAScanner } from './scanners/EAScanner.js';
+import { BattleNetScanner } from './scanners/BattleNetScanner.js';
+import { HumbleScanner } from './scanners/HumbleScanner.js';
+import { ItchScanner } from './scanners/ItchScanner.js';
+import { ManualScanner } from './scanners/ManualScanner.js';
+import { DeepScanner } from './scanners/DeepScanner.js';
+
+// Re-export ScannedGameResult for compatibility
+export type { ScannedGameResult } from './scanners/ScannedGameResult.js';
 
 /**
  * Service to orchestrate scanning from all game sources
@@ -38,6 +35,19 @@ export class ImportService {
   private gameFilteringService: GameFilteringService;
   private isScanCancelled: boolean = false;
 
+  // Scanners
+  private steamScanner: SteamScanner;
+  private xboxScanner: XboxScanner;
+  private epicScanner: EpicScanner;
+  private gogScanner: GogScanner;
+  private ubisoftScanner: UbisoftScanner;
+  private rockstarScanner: RockstarScanner;
+  private eaScanner: EAScanner;
+  private battleNetScanner: BattleNetScanner;
+  private humbleScanner: HumbleScanner;
+  private itchScanner: ItchScanner;
+  private deepScanner: DeepScanner;
+
   constructor(
     steamService: SteamService,
     xboxService: XboxService,
@@ -49,6 +59,19 @@ export class ImportService {
     this.appConfigService = appConfigService;
     this.metadataFetcher = metadataFetcher;
     this.gameFilteringService = new GameFilteringService();
+
+    // Initialize Scanners
+    this.steamScanner = new SteamScanner(this.steamService);
+    this.xboxScanner = new XboxScanner(this.xboxService);
+    this.epicScanner = new EpicScanner(this.gameFilteringService);
+    this.gogScanner = new GogScanner(this.gameFilteringService);
+    this.ubisoftScanner = new UbisoftScanner(this.gameFilteringService);
+    this.rockstarScanner = new RockstarScanner(this.gameFilteringService);
+    this.eaScanner = new EAScanner(this.gameFilteringService);
+    this.battleNetScanner = new BattleNetScanner(this.gameFilteringService);
+    this.humbleScanner = new HumbleScanner(this.gameFilteringService);
+    this.itchScanner = new ItchScanner(this.gameFilteringService);
+    this.deepScanner = new DeepScanner(this.gameFilteringService);
   }
 
   /**
@@ -114,30 +137,43 @@ export class ImportService {
           console.log(`[ImportService] Scanning ${config.id} at path: ${config.path}`);
 
           let games: ScannedGameResult[] = [];
-          if (config.id === 'steam') {
-            games = await this.scanSteam(config.path);
-          } else if (config.id === 'xbox') {
-            games = await this.scanXbox(config.path);
-          } else if (config.id === 'epic') {
-            games = await this.scanEpic(config.path);
-          } else if (config.id === 'gog') {
-            games = await this.scanGOG(config.path);
-          } else if (config.id === 'ubisoft') {
-            games = await this.scanUbisoft(config.path);
-          } else if (config.id === 'rockstar') {
-            games = await this.scanRockstar(config.path);
-          } else if (config.id === 'ea') {
-            games = await this.scanEA(config.path);
-          } else if (config.id === 'battle') {
-            games = await this.scanBattle(config.path);
-          } else if (config.id === 'humble') {
-            games = await this.scanHumble(config.path);
-          } else if (config.id === 'itch') {
-            games = await this.scanItch(config.path);
-          } else {
-            // Fallback: Use generic deep scan for any unknown app type
-            console.log(`[ImportService] No specific scanner for ${config.id}, using generic deep scan`);
-            games = await this.scanFolderForExecutables(config.path);
+
+          switch (config.id) {
+            case 'steam':
+              games = await this.steamScanner.scan(config.path);
+              break;
+            case 'xbox':
+              games = await this.xboxScanner.scan(config.path);
+              break;
+            case 'epic':
+              games = await this.epicScanner.scan(config.path);
+              break;
+            case 'gog':
+              games = await this.gogScanner.scan(config.path);
+              break;
+            case 'ubisoft':
+              games = await this.ubisoftScanner.scan(config.path);
+              break;
+            case 'rockstar':
+              games = await this.rockstarScanner.scan(config.path);
+              break;
+            case 'ea':
+              games = await this.eaScanner.scan(config.path);
+              break;
+            case 'battle':
+              games = await this.battleNetScanner.scan(config.path);
+              break;
+            case 'humble':
+              games = await this.humbleScanner.scan(config.path);
+              break;
+            case 'itch':
+              games = await this.itchScanner.scan(config.path);
+              break;
+            default:
+              // Fallback: Use generic deep scan for any unknown app type
+              console.log(`[ImportService] No specific scanner for ${config.id}, using generic deep scan`);
+              games = await this.deepScanner.scan(config.path);
+              break;
           }
 
           if (games.length > 0) {
@@ -173,7 +209,10 @@ export class ImportService {
           try {
             if (existsSync(folder)) {
               progressCallback?.(`Scanning manual root ${folder} (subfolders = game names)...`);
-              const folderGames = this.scanManualFolder(folder, config.autoCategory, config.name);
+
+              const manualScanner = new ManualScanner(config.autoCategory, config.name, this.gameFilteringService);
+              const folderGames = await manualScanner.scan(folder);
+
               if (folderGames.length > 0) {
                 progressCallback?.(`Found ${folderGames.length} game${folderGames.length !== 1 ? 's' : ''} in ${folder}`);
                 folderGames.forEach(game => {
@@ -201,7 +240,7 @@ export class ImportService {
         try {
           progressCallback?.('Auto-detecting Battle.net games...');
           console.log('[ImportService] Battle.net not configured, running auto-detection via registry...');
-          const battleGames = await this.scanBattle('');
+          const battleGames = await this.battleNetScanner.scan('');
           if (battleGames.length > 0) {
             progressCallback?.(`Found ${battleGames.length} Battle.net game${battleGames.length !== 1 ? 's' : ''} (auto-detected)`);
             battleGames.forEach(game => {
@@ -218,1656 +257,6 @@ export class ImportService {
     } catch (error) {
       console.error('Error in scanAllSources:', error);
       progressCallback?.(`Error: ${error instanceof Error ? error.message : 'Unknown error'}`);
-    }
-
-    return results;
-  }
-
-  /**
-   * Scan Steam games
-   */
-  private async scanSteam(steamPath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[ImportService] scanSteam called with path: ${steamPath}`);
-      this.steamService.setSteamPath(steamPath);
-      const steamGames = this.steamService.scanSteamGames();
-      console.log(`[ImportService] scanSteam found ${steamGames.length} games`);
-
-      return steamGames.map((game: SteamGame) => {
-        const categories: string[] = [];
-        if (game.name.toLowerCase().includes('steamvr')) {
-          categories.push('VR');
-        }
-
-        return {
-          uuid: `steam-${game.appId}-${Date.now()}`,
-          source: 'steam' as const,
-          originalName: game.name,
-          installPath: game.installDir.replace(/\\/g, '/'),
-          exePath: undefined, // Steam games use steam:// protocol
-          appId: game.appId,
-          title: game.name,
-          categories,
-          status: 'ready' as const, // Steam games with AppID are ready
-          isDownloading: false, // Steam Service already filters for installed games
-        };
-      });
-    } catch (error) {
-      console.error('Error scanning Steam:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan Xbox games
-   */
-  private async scanXbox(xboxPath: string): Promise<ScannedGameResult[]> {
-    try {
-      const xboxGames = this.xboxService.scanGames(xboxPath);
-
-      return xboxGames.map((game: XboxGame) => {
-        // XboxService returns installPath as the full exe path for both UWP and PC games
-        // Extract the folder path for installPath, keep exe path separate
-        const exePath = game.type === 'pc' ? game.installPath : undefined;
-        let installPath = game.installPath;
-        if (game.installPath) {
-          const pathParts = game.installPath.split(/[/\\]/);
-          if (pathParts.length > 1) {
-            pathParts.pop(); // Remove the exe filename when present
-            installPath = pathParts.join(sep);
-          }
-        }
-
-        return {
-          uuid: `${game.id}-${Date.now()}`,
-          source: 'xbox' as const,
-          originalName: game.name,
-          installPath: installPath.replace(/\\/g, '/'),
-          exePath: exePath?.replace(/\\/g, '/'),
-          appId: game.appId,
-          packageFamilyName: game.packageFamilyName,
-          appUserModelId: game.appUserModelId,
-          launchUri: game.launchUri,
-          xboxKind: game.type,
-          title: game.name,
-          status: 'ambiguous' as const, // Xbox games need metadata matching
-          isDownloading: this.gameFilteringService.isLikelyDownloading(installPath),
-        };
-      });
-    } catch (error) {
-      console.error('Error scanning Xbox:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan Epic Games Launcher for installed games
-   * First tries to read from manifests, then falls back to scanning game folders directly
-   */
-  private async scanEpic(epicPath: string): Promise<ScannedGameResult[]> {
-    try {
-      const results: ScannedGameResult[] = [];
-
-      // Epic Games stores manifests typically in ProgramData; also try under provided epicPath
-      const defaultManifests = join(process.env.ProgramData || 'C:\\ProgramData', 'Epic', 'EpicGamesLauncher', 'Data', 'Manifests');
-      const manifestsPath = existsSync(defaultManifests)
-        ? defaultManifests
-        : join(epicPath, 'Epic Games Launcher', 'Data', 'Manifests');
-
-      if (existsSync(manifestsPath)) {
-        // Try to read from manifests first (preferred method)
-        const manifestFiles = readdirSync(manifestsPath).filter(f => f.endsWith('.item'));
-
-        for (const manifestFile of manifestFiles) {
-          try {
-            const manifestPath = join(manifestsPath, manifestFile);
-            const manifestContent = readFileSync(manifestPath, 'utf-8');
-            const manifest = JSON.parse(manifestContent);
-
-            // Epic manifest structure
-            const appName = manifest.DisplayName || manifest.LaunchExecutable || 'Unknown';
-            const installLocation = manifest.InstallLocation;
-            const launchExecutable = manifest.LaunchExecutable;
-            const catalogNamespace = manifest.CatalogNamespace;
-            const catalogItemId = manifest.CatalogItemId;
-
-            if (installLocation && existsSync(installLocation)) {
-              // Find the executable
-              let exePath: string | undefined;
-
-              if (launchExecutable) {
-                const fullExePath = join(installLocation, launchExecutable);
-                if (existsSync(fullExePath)) {
-                  exePath = fullExePath;
-                }
-              }
-
-              // If no launch executable specified, search for common exe names
-              if (!exePath) {
-                const commonExes = [
-                  join(installLocation, `${appName}.exe`),
-                  join(installLocation, 'Binaries', 'Win64', `${appName}.exe`),
-                  join(installLocation, 'Binaries', 'Win32', `${appName}.exe`),
-                ];
-
-                for (const exe of commonExes) {
-                  if (existsSync(exe)) {
-                    exePath = exe;
-                    break;
-                  }
-                }
-              }
-
-              // Create a unique ID from catalog info or use manifest filename
-              const appId = catalogItemId || manifestFile.replace('.item', '');
-
-              results.push({
-                uuid: `epic-${appId}-${Date.now()}`,
-                source: 'epic' as const,
-                originalName: appName,
-                installPath: installLocation.replace(/\\/g, '/'),
-                exePath: exePath?.replace(/\\/g, '/'),
-                appId: appId,
-                title: appName,
-                status: 'ambiguous' as const, // Epic games need metadata matching
-                isDownloading: this.gameFilteringService.isLikelyDownloading(installLocation),
-              });
-            }
-          } catch (err) {
-            console.error(`Error parsing Epic manifest ${manifestFile}:`, err);
-            continue;
-          }
-        }
-      }
-
-      // Fallback: Scan Epic Games directory directly for game folders
-      // This handles cases where manifests aren't available or games are installed directly
-      if (!existsSync(epicPath)) {
-        return results;
-      }
-
-      try {
-        const entries = readdirSync(epicPath);
-
-        for (const entry of entries) {
-          // Skip the Epic Games Launcher folder and other non-game folders
-          if (entry === 'Epic Games Launcher' || entry === 'UnrealEngine') {
-            continue;
-          }
-
-          const gamePath = join(epicPath, entry);
-
-          try {
-            const stats = statSync(gamePath);
-            if (!stats.isDirectory()) {
-              continue;
-            }
-
-            // Check if we already found this game from manifests
-            const alreadyFound = results.some(r => r.installPath === gamePath);
-            if (alreadyFound) {
-              continue;
-            }
-
-            // Look for executables in this game folder (deep scan)
-            const exeFiles = this.findExecutables(gamePath, 0, 20);
-
-            // Filter out helper executables
-            const gameExes = exeFiles.filter(exe => {
-              const fileName = exe.toLowerCase();
-              const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-              return !fileNameOnly.includes('gamelaunchhelper') &&
-                !fileNameOnly.includes('bootstrapper') &&
-                !fileNameOnly.includes('uninstall') &&
-                !fileNameOnly.includes('setup') &&
-                !fileNameOnly.includes('installer') &&
-                !fileNameOnly.includes('launcher') &&
-                fileNameOnly !== 'crashreportclient.exe' &&
-                fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-                fileNameOnly !== 'battlenet.exe' &&
-                fileNameOnly !== 'battle.net.exe' &&
-                fileNameOnly !== 'crashpad_handler.exe' &&
-                fileNameOnly !== 'embark-crash-helper.exe' &&
-                fileNameOnly !== 'blizzardbrowser.exe' &&
-                fileNameOnly !== 'blizzarderror.exe' &&
-                fileNameOnly !== 'gamesessionmonitor.exe';
-            });
-
-            if (gameExes.length > 0) {
-              // Use the first executable found
-              const mainExe = gameExes[0];
-
-              results.push({
-                uuid: `epic-${entry}-${Date.now()}`,
-                source: 'epic' as const,
-                originalName: entry,
-                installPath: gamePath.replace(/\\/g, '/'),
-                exePath: mainExe.replace(/\\/g, '/'),
-                appId: undefined,
-                title: entry,
-                status: 'ambiguous' as const, // Epic games need metadata matching
-                isDownloading: this.gameFilteringService.isLikelyDownloading(gamePath),
-              });
-            }
-          } catch (err) {
-            // Skip folders we can't access
-            continue;
-          }
-        }
-      } catch (err) {
-        console.error('Error scanning Epic Games directory:', err);
-      }
-
-      return results;
-    } catch (error) {
-      console.error('Error scanning Epic Games:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan GOG Galaxy for installed games
-   * Games are located in: {GOGPath}\Games
-   * Example: C:\Program Files (x86)\GOG Galaxy\Games
-   */
-  private async scanGOG(gogPath: string): Promise<ScannedGameResult[]> {
-    try {
-      const results: ScannedGameResult[] = [];
-
-      // GOG games are in: {GOGPath}\Games
-      // If path is already the Games folder, use it directly
-      let gamesPath: string;
-      if (gogPath.toLowerCase().endsWith('games')) {
-        gamesPath = gogPath;
-      } else {
-        gamesPath = join(gogPath, 'Games');
-      }
-
-      if (!existsSync(gamesPath)) {
-        // Try alternative location
-        const altGamesPath = join(gogPath, 'Galaxy', 'Games');
-        if (existsSync(altGamesPath)) {
-          return this.scanGOGGamesFolder(altGamesPath);
-        }
-        console.warn(`GOG Games folder not found: ${gamesPath}`);
-        return results;
-      }
-
-      return this.scanGOGGamesFolder(gamesPath);
-    } catch (error) {
-      console.error('Error scanning GOG:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan GOG games folder for executables
-   * Recursively scans all subdirectories and creates a game entry for each folder containing a valid executable
-   */
-  private scanGOGGamesFolder(gamesPath: string): ScannedGameResult[] {
-    const results: ScannedGameResult[] = [];
-
-    try {
-      console.log(`[GOG] Scanning games folder: ${gamesPath}`);
-
-      // Find all executables recursively in the folder
-      const allExeFiles = this.findExecutables(gamesPath, 0, 20);
-      console.log(`[GOG] Found ${allExeFiles.length} total executables`);
-
-      // Filter out helper executables
-      const gameExes = allExeFiles.filter(exe => {
-        const fileName = exe.toLowerCase();
-        const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-        return !fileNameOnly.includes('gamelaunchhelper') &&
-          !fileNameOnly.includes('bootstrapper') &&
-          !fileNameOnly.includes('uninstall') &&
-          !fileNameOnly.includes('setup') &&
-          !fileNameOnly.includes('installer') &&
-          fileNameOnly !== 'crashreportclient.exe' &&
-          fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-          fileNameOnly !== 'battlenet.exe' &&
-          fileNameOnly !== 'battle.net.exe' &&
-          fileNameOnly !== 'cleanup.exe' &&
-          fileNameOnly !== 'crashpad_handler.exe' &&
-          fileNameOnly !== 'gamesessionmonitor.exe';
-      });
-
-      console.log(`[GOG] Filtered to ${gameExes.length} valid game executables`);
-
-      if (gameExes.length === 0) {
-        return results;
-      }
-
-      // Group executables by their containing directory
-      const gamesByFolder = new Map<string, string[]>();
-
-      for (const exePath of gameExes) {
-        const exeDir = dirname(exePath);
-        const normalizedDir = exeDir.replace(/\\/g, sep).replace(/\/\//g, '/');
-
-        if (!gamesByFolder.has(normalizedDir)) {
-          gamesByFolder.set(normalizedDir, []);
-        }
-        gamesByFolder.get(normalizedDir)!.push(exePath);
-      }
-
-      console.log(`[GOG] Found ${gamesByFolder.size} unique game folders`);
-
-      // Create a game entry for each folder
-      for (const [gameDir, exePaths] of gamesByFolder.entries()) {
-        try {
-          // Only process directories within the games path
-          const normalizedGamesPath = gamesPath.replace(/\\/g, sep).replace(/\/\//g, '/');
-          if (!gameDir.startsWith(normalizedGamesPath) && gameDir !== normalizedGamesPath) {
-            continue;
-          }
-
-          // Use the folder name as the default game title
-          const folderName = gameDir.split(sep).pop() || 'Unknown';
-          let gameTitle = folderName;
-
-          // For common launcher/emulator subfolders (scummvm, dosbox, etc.), use parent folder name
-          const commonEmulators = ['scummvm', 'dosbox', 'dosboxstaging', 'wine', 'proton'];
-          if (commonEmulators.includes(folderName.toLowerCase())) {
-            // Get parent folder name instead
-            const parentFolder = gameDir.split(sep).slice(0, -1).pop();
-            if (parentFolder && parentFolder.toLowerCase() !== 'games') {
-              gameTitle = parentFolder;
-              console.log(`[GOG] Using parent folder name for emulator subfolder: ${gameTitle}`);
-            }
-          }
-
-          // Try to read GOG Galaxy .info file for game metadata
-          let launchArgs: string | undefined;
-          let primaryExePath: string | undefined;
-          let infoData: any;
-          let infoFile: string | undefined;
-          let infoDir: string = gameDir;
-
-          // Search up to the GOG games root for an .info file
-          let searchPath = gameDir;
-          // Reuse normalizedGamesPath from above
-
-          while (searchPath.startsWith(normalizedGamesPath)) {
-            try {
-              const contents = readdirSync(searchPath);
-              const found = contents.find(f => f.match(/^goggame-\d+\.info$/));
-              if (found) {
-                infoFile = found;
-                infoDir = searchPath;
-                break;
-              }
-            } catch (err) {
-              // Ignore access errors
-            }
-
-            const parent = dirname(searchPath);
-            if (parent === searchPath) break;
-            searchPath = parent;
-          }
-
-          if (infoFile) {
-            try {
-              const infoPath = join(infoDir, infoFile);
-              const infoContent = readFileSync(infoPath, 'utf-8');
-              infoData = JSON.parse(infoContent);
-
-              // Extract the game name from .info file if available
-              if (infoData.name) {
-                gameTitle = infoData.name;
-              }
-
-              // Find the primary game task
-              const playTasks = infoData.playTasks || [];
-              const primaryTask = playTasks.find((task: any) => task.isPrimary && task.category === 'game');
-
-              if (primaryTask && primaryTask.path) {
-                // Resolve the relative path to absolute, relative to where the .info file is
-                const taskPath = primaryTask.path.replace(/\\/g, sep);
-                const absolutePath = join(infoDir, taskPath);
-
-                // Check if the resolved exe exists
-                if (existsSync(absolutePath)) {
-                  primaryExePath = absolutePath;
-                }
-
-                // Extract launch arguments
-                if (primaryTask.arguments) {
-                  launchArgs = primaryTask.arguments;
-                  console.log(`[GOG] Found launch arguments for ${gameTitle}: ${launchArgs}`);
-                }
-              }
-            } catch (err) {
-              console.warn(`[GOG] Could not parse info file at ${infoDir}:`, err);
-            }
-          }
-
-          // Select the best executable from this folder
-          let mainExe = primaryExePath || exePaths[0];
-
-          // If no primary exe from info, try to match by folder name
-          if (!primaryExePath) {
-            const matchingExe = exePaths.find(exe => {
-              const exeName = exe.split(sep).pop()?.toLowerCase().replace('.exe', '') || '';
-              return exeName === folderName.toLowerCase();
-            });
-
-            if (matchingExe) {
-              mainExe = matchingExe;
-            } else {
-              // Prefer executables closer to the folder root
-              const sortedByDepth = exePaths.sort((a, b) => {
-                const depthA = a.substring(gameDir.length).split(sep).length;
-                const depthB = b.substring(gameDir.length).split(sep).length;
-                return depthA - depthB;
-              });
-              mainExe = sortedByDepth[0];
-            }
-          }
-
-          results.push({
-            uuid: `gog-${gameDir}-${Date.now()}`,
-            source: 'gog' as const,
-            originalName: gameTitle,
-            installPath: gameDir.replace(/\\/g, '/'),
-            exePath: mainExe?.replace(/\\/g, '/'),
-            launchArgs: launchArgs,
-            appId: infoFile ? infoFile.replace(/\D+/g, '') : undefined,
-            title: gameTitle,
-            status: 'ambiguous' as const, // GOG games need metadata matching
-            isDownloading: this.gameFilteringService.isLikelyDownloading(gameDir),
-          });
-        } catch (err) {
-          console.warn(`[GOG] Could not process game folder "${gameDir}":`, err);
-          continue;
-        }
-      }
-
-      console.log(`[GOG] Found ${results.length} games total`);
-    } catch (err) {
-      console.error('[GOG] Error scanning GOG games folder:', err);
-    }
-
-    return results;
-  }
-
-  /**
-   * Scan manual root folder where each immediate subfolder is treated as a game.
-   * Game title = subfolder name. Picks the first viable executable under that subfolder.
-   */
-  private scanManualFolder(rootPath: string, autoCategory?: string[], folderName?: string): ScannedGameResult[] {
-    const results: ScannedGameResult[] = [];
-
-    try {
-      const entries = readdirSync(rootPath);
-
-      for (const entry of entries) {
-        const gameDir = join(rootPath, entry);
-        try {
-          const stats = statSync(gameDir);
-          if (!stats.isDirectory()) {
-            continue;
-          }
-
-          const title = entry;
-
-          // Find executables within the game folder (allow deeper nests, but reuse global filters)
-          const exeCandidates = this.findExecutables(gameDir, 0, 20);
-
-          // Choose best executable: prefer filename matching folder name, otherwise closest to root
-          let mainExe: string | undefined;
-          if (exeCandidates.length > 0) {
-            const normalizedTitle = title.toLowerCase();
-            const exactMatch = exeCandidates.find(p => {
-              const exeName = p.split(/[/\\]/).pop()?.toLowerCase().replace('.exe', '') || '';
-              return exeName === normalizedTitle;
-            });
-
-            if (exactMatch) {
-              mainExe = exactMatch;
-            } else {
-              // Pick the exe with the shortest relative path (closest to root)
-              mainExe = exeCandidates.sort((a, b) => {
-                const relA = a.substring(gameDir.length).split(sep).length;
-                const relB = b.substring(gameDir.length).split(sep).length;
-                return relA - relB;
-              })[0];
-            }
-          }
-
-          // Try to detect GOG metadata for manual folders
-          let launchArgs: string | undefined;
-          let appId: string | undefined;
-          let finalTitle = title;
-          let finalExe = mainExe;
-
-          try {
-            const dirContents = readdirSync(gameDir);
-            const infoFile = dirContents.find(f => f.match(/^goggame-(\d+)\.info$/));
-            if (infoFile) {
-              const infoPath = join(gameDir, infoFile);
-              const infoContent = readFileSync(infoPath, 'utf-8');
-              const infoData = JSON.parse(infoContent);
-
-              if (infoData.name) finalTitle = infoData.name;
-              appId = infoFile.replace(/\D+/g, '');
-
-              const playTasks = infoData.playTasks || [];
-              const primaryTask = playTasks.find((task: any) => task.isPrimary && task.category === 'game');
-              if (primaryTask) {
-                if (primaryTask.path) {
-                  const taskPath = primaryTask.path.replace(/\\/g, sep);
-                  const absolutePath = join(gameDir, taskPath);
-                  if (existsSync(absolutePath)) {
-                    finalExe = absolutePath;
-                  }
-                }
-                if (primaryTask.arguments) {
-                  launchArgs = primaryTask.arguments;
-                  console.log(`[Manual-GOG] Detected launch arguments for ${finalTitle}: ${launchArgs}`);
-                }
-              }
-            }
-          } catch (err) {
-            // Ignore metadata detection errors
-          }
-
-          results.push({
-            uuid: `manual-${gameDir}-${Date.now()}`,
-            source: folderName || 'Manual Folder',
-            originalName: title,
-            installPath: gameDir.replace(/\\/g, '/'),
-            exePath: finalExe?.replace(/\\/g, '/'),
-            launchArgs: launchArgs,
-            appId: appId,
-            title: finalTitle,
-            categories: autoCategory,
-            status: 'ambiguous' as const,
-          });
-        } catch (err) {
-          console.warn(`[Manual] Could not process manual subfolder "${gameDir}":`, err);
-          continue;
-        }
-      }
-    } catch (error) {
-      console.error('[Manual] Error scanning manual root:', error);
-    }
-
-    return results;
-  }
-
-  /**
-   * Find executable files in a directory (recursive, deep scan with high max depth)
-   */
-  private findExecutables(dirPath: string, depth: number = 0, maxDepth: number = 20): string[] {
-    const executables: string[] = [];
-
-    if (depth > maxDepth) return executables;
-
-    const excludeNames = [
-      'gamelaunchhelper.exe',
-      'bootstrapper.exe',
-      'crashreportclient.exe',
-      'battlenet.overlay.runtime.exe',
-      'battlenet.exe',
-      'battle.net.exe',
-      'crashpad_handler.exe',
-      'embark-crash-helper.exe',
-      'blizzardbrowser.exe',
-      'blizzarderror.exe',
-      'gamesessionmonitor.exe',
-      'unins000.exe',
-      'autorun.exe',
-      'vc_redist.x64.exe',
-      'gamelaunchhelper',
-      'bootstrapper',
-      'crashreportclient',
-      'battlenet.overlay.runtime',
-      'crashpad_handler',
-      'embark-crash-helper',
-      'blizzardbrowser',
-      'blizzarderror',
-      'gamesessionmonitor',
-      'unins000',
-      'autorun',
-      'vc_redist.x64',
-    ];
-
-    try {
-      const entries = readdirSync(dirPath);
-
-      for (const entry of entries) {
-        const fullPath = join(dirPath, entry);
-
-        try {
-          const stats = statSync(fullPath);
-
-          if (stats.isFile() && entry.toLowerCase().endsWith('.exe')) {
-            const lowerName = entry.toLowerCase();
-            const baseName = lowerName.replace('.exe', '');
-
-            // Check exact matches first
-            if (excludeNames.includes(lowerName) || excludeNames.includes(baseName)) {
-              continue;
-            }
-
-            // Check patterns - be aggressive about filtering non-game executables
-            // Optimized local import approach
-            if (!lowerName.includes('installer') &&
-              !lowerName.includes('setup') &&
-              !lowerName.includes('uninstall') &&
-              !lowerName.includes('unins') &&
-              !lowerName.includes('autorun') &&
-              !lowerName.includes('vc_redist') &&
-              !lowerName.includes('vcredist') &&
-              !lowerName.includes('launcher') &&
-              !lowerName.includes('updater') &&
-              !lowerName.includes('update') &&
-              !lowerName.includes('patcher') &&
-              !lowerName.includes('repair') &&
-              !lowerName.includes('config') &&
-              !lowerName.includes('settings') &&
-              !lowerName.includes('benchmark') &&
-              !lowerName.includes('diagnostic') &&
-              !lowerName.includes('reporter') &&
-              !lowerName.includes('monitor') &&
-              !lowerName.includes('helper') &&
-              !lowerName.includes('service') &&
-              !lowerName.includes('daemon') &&
-              !lowerName.includes('agent') &&
-              !lowerName.includes('overlay') &&
-              !lowerName.includes('gamelaunchhelper') &&
-              !lowerName.includes('bootstrapper') &&
-              !lowerName.includes('crashreport') &&
-              !lowerName.includes('crash_report') &&
-              !lowerName.includes('crashhandler') &&
-              !lowerName.includes('crash_handler') &&
-              !lowerName.includes('errorhandler') &&
-              !lowerName.includes('battlenet.overlay.runtime') &&
-              !lowerName.includes('crashpad') &&
-              !lowerName.includes('embark-crash') &&
-              !lowerName.includes('blizzardbrowser') &&
-              !lowerName.includes('blizzarderror') &&
-              !lowerName.includes('gamesessionmonitor') &&
-              !lowerName.includes('uplay') &&
-              !lowerName.includes('ubisoft') &&
-              !lowerName.includes('eadesktop') &&
-              !lowerName.includes('origin') &&
-              !lowerName.includes('epicgames') &&
-              !lowerName.includes('steam') &&
-              !lowerName.includes('activation') &&
-              !lowerName.includes('redist') &&
-              !lowerName.includes('directx') &&
-              !lowerName.includes('prereq') &&
-              !lowerName.includes('_redist') &&
-              !baseName.startsWith('ue4') &&
-              !baseName.startsWith('ue5') &&
-              !baseName.endsWith('server') &&
-              !baseName.endsWith('editor')) {
-              executables.push(fullPath);
-            }
-          } else if (stats.isDirectory() && depth < maxDepth) {
-            // Skip folders that definitely don't contain primary game executables
-            const dirName = entry.toLowerCase();
-            if (dirName.includes('wingdk') ||
-              dirName === 'update' ||
-              dirName === 'updater' ||
-              dirName === 'setup' ||
-              dirName === 'install' ||
-              dirName === 'installer' ||
-              dirName === 'redist' ||
-              dirName === 'redistributables' ||
-              dirName === 'directx' ||
-              dirName === 'dotnet' ||
-              dirName === '_commonredist' ||
-              dirName === 'support' ||
-              dirName === 'engine' ||
-              dirName === 'binaries' || // Some games use Binaries, but often primary exe is at root
-              dirName === 'tools' ||
-              dirName === 'benchmark' ||
-              dirName === 'extras' ||
-              dirName === 'bonus' ||
-              dirName === 'soundtrack' ||
-              dirName === 'artbook' ||
-              dirName === 'manuals' ||
-              dirName === 'patch' ||
-              dirName === 'patches') {
-              continue;
-            }
-            // Recursively search subdirectories
-            const subExes = this.findExecutables(fullPath, depth + 1, maxDepth);
-            executables.push(...subExes);
-          }
-        } catch (err) {
-          // Skip entries we can't access
-          continue;
-        }
-      }
-    } catch (err) {
-      // Skip directories we can't access
-    }
-
-    return executables;
-  }
-
-  /**
-   * Scan Ubisoft Connect for installed games
-   * Games are located in: {UbisoftPath}\games
-   * Example: C:\Program Files (x86)\Ubisoft\Ubisoft Game Launcher\games
-   */
-  private async scanUbisoft(ubisoftPath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[Ubisoft] Starting scan with path: ${ubisoftPath}`);
-      const results: ScannedGameResult[] = [];
-
-      // Ubisoft games are in: {UbisoftPath}\games (lowercase)
-      // If path is already the games folder, use it directly
-      let gamesPath: string;
-      if (ubisoftPath.toLowerCase().endsWith('games')) {
-        gamesPath = ubisoftPath;
-      } else {
-        gamesPath = join(ubisoftPath, 'games');
-      }
-
-      console.log(`[Ubisoft] Checking games folder: ${gamesPath}`);
-
-      if (!existsSync(gamesPath)) {
-        console.warn(`[Ubisoft] Games folder not found: ${gamesPath}`);
-        // Try alternative path structure (some installations might be different)
-        const altPath = join(ubisoftPath, 'Games');
-        if (existsSync(altPath)) {
-          console.log(`[Ubisoft] Found games in alternative path: ${altPath}`);
-          gamesPath = altPath;
-        } else {
-          return results;
-        }
-      }
-
-      console.log(`[Ubisoft] Scanning games folder: ${gamesPath}`);
-      const scannedGames = this.scanUbisoftGamesFolder(gamesPath);
-      console.log(`[Ubisoft] Found ${scannedGames.length} games`);
-      return scannedGames;
-    } catch (error) {
-      console.error('[Ubisoft] Error scanning Ubisoft:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan Ubisoft games folder for executables
-   * Recursively scans all subdirectories and creates a game entry for each folder containing a valid executable
-   */
-  private scanUbisoftGamesFolder(gamesPath: string): ScannedGameResult[] {
-    const results: ScannedGameResult[] = [];
-
-    try {
-      console.log(`[Ubisoft] Scanning games folder: ${gamesPath}`);
-
-      // Find all executables recursively in the folder
-      const allExeFiles = this.findExecutables(gamesPath, 0, 20);
-      console.log(`[Ubisoft] Found ${allExeFiles.length} total executables`);
-
-      // Filter out helper executables
-      const gameExes = allExeFiles.filter(exe => {
-        const fileName = exe.toLowerCase();
-        const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-        return !fileNameOnly.includes('gamelaunchhelper') &&
-          !fileNameOnly.includes('bootstrapper') &&
-          !fileNameOnly.includes('uninstall') &&
-          !fileNameOnly.includes('setup') &&
-          !fileNameOnly.includes('installer') &&
-          !fileNameOnly.includes('uplay') &&
-          !fileNameOnly.includes('ubisoft') &&
-          fileNameOnly !== 'crashreportclient.exe' &&
-          fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-          fileNameOnly !== 'battlenet.exe' &&
-          fileNameOnly !== 'battle.net.exe' &&
-          fileNameOnly !== 'cleanup.exe' &&
-          fileNameOnly !== 'crashpad_handler.exe' &&
-          fileNameOnly !== 'gamesessionmonitor.exe';
-      });
-
-      console.log(`[Ubisoft] Filtered to ${gameExes.length} valid game executables`);
-
-      if (gameExes.length === 0) {
-        return results;
-      }
-
-      // Group executables by their containing directory
-      const gamesByFolder = new Map<string, string[]>();
-
-      for (const exePath of gameExes) {
-        const exeDir = dirname(exePath);
-        const normalizedDir = exeDir.replace(/\\/g, sep).replace(/\/\//g, '/');
-
-        if (!gamesByFolder.has(normalizedDir)) {
-          gamesByFolder.set(normalizedDir, []);
-        }
-        gamesByFolder.get(normalizedDir)!.push(exePath);
-      }
-
-      console.log(`[Ubisoft] Found ${gamesByFolder.size} unique game folders`);
-
-      // Create a game entry for each folder
-      for (const [gameDir, exePaths] of gamesByFolder.entries()) {
-        try {
-          // Only process directories within the games path
-          const normalizedGamesPath = gamesPath.replace(/\\/g, sep).replace(/\/\//g, '/');
-          if (!gameDir.startsWith(normalizedGamesPath) && gameDir !== normalizedGamesPath) {
-            continue;
-          }
-
-          // Use the folder name as the game title
-          let folderName = gameDir.split(sep).pop() || 'Unknown';
-
-          // Detect Ubisoft download/staging structure
-          if (gameDir.toLowerCase().includes('uplay_download')) {
-            const parts = gameDir.split(/[/\\]/);
-            const downloadIdx = parts.findIndex(p => p.toLowerCase() === 'uplay_download');
-            if (downloadIdx > 0) {
-              // The parent of 'uplay_download' is usually the actual game name
-              folderName = parts[downloadIdx - 1];
-              console.log(`[Ubisoft] Detected download folder, using real game name from parent: ${folderName}`);
-            }
-          }
-
-          // Select the best executable from this folder
-          let mainExe = exePaths[0];
-
-          // Prefer executables with the same name as the directory (or the parent directory we just found)
-          const matchingExe = exePaths.find(exe => {
-            const exeName = exe.split(sep).pop()?.toLowerCase().replace('.exe', '') || '';
-            const normalizedFolder = folderName.toLowerCase();
-            return exeName === normalizedFolder || normalizedFolder.includes(exeName);
-          });
-
-          if (matchingExe) {
-            mainExe = matchingExe;
-          } else {
-            // Prefer executables closer to the folder root
-            const sortedByDepth = exePaths.sort((a, b) => {
-              const depthA = a.substring(gameDir.length).split(sep).length;
-              const depthB = b.substring(gameDir.length).split(sep).length;
-              return depthA - depthB;
-            });
-            mainExe = sortedByDepth[0];
-          }
-
-          console.log(`[Ubisoft] Adding game: ${folderName} with exe: ${mainExe}`);
-
-          results.push({
-            uuid: `ubisoft-${gameDir}-${Date.now()}`,
-            source: 'ubisoft' as const,
-            originalName: folderName,
-            installPath: gameDir,
-            exePath: mainExe,
-            appId: undefined,
-            title: folderName,
-            status: 'ambiguous' as const, // Ubisoft games need metadata matching
-            isDownloading: this.gameFilteringService.isLikelyDownloading(gameDir),
-          });
-        } catch (err) {
-          console.warn(`[Ubisoft] Could not process game folder "${gameDir}":`, err);
-          continue;
-        }
-      }
-
-      console.log(`[Ubisoft] Found ${results.length} games total`);
-    } catch (err) {
-      console.error('[Ubisoft] Error scanning Ubisoft games folder:', err);
-    }
-
-    return results;
-  }
-
-  /**
-   * Scan Rockstar Games Launcher for installed games
-   * Games are typically in: {RockstarPath}\{GameName}
-   * Example: C:\Program Files\Rockstar Games\Grand Theft Auto V
-   */
-  private async scanRockstar(rockstarPath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[Rockstar] Starting scan with path: ${rockstarPath}`);
-      const results: ScannedGameResult[] = [];
-
-      if (!existsSync(rockstarPath)) {
-        console.warn(`[Rockstar] Rockstar Games folder not found: ${rockstarPath}`);
-        return results;
-      }
-
-      // Rockstar games are typically in subdirectories of the Rockstar Games folder
-      // Each game has its own folder (e.g., Grand Theft Auto V, Red Dead Redemption 2)
-      try {
-        const entries = readdirSync(rockstarPath);
-        console.log(`[Rockstar] Found ${entries.length} entries in Rockstar Games folder`);
-
-        for (const entry of entries) {
-          // Skip known non-game folders
-          const dirName = entry.toLowerCase();
-          if (dirName === 'launcher' ||
-            dirName === 'social club' ||
-            dirName === 'redistributables' ||
-            dirName.startsWith('$')) {
-            console.log(`[Rockstar] Skipping non-game folder: ${entry}`);
-            continue;
-          }
-
-          const gamePath = join(rockstarPath, entry);
-
-          try {
-            const stats = statSync(gamePath);
-            if (!stats.isDirectory()) {
-              continue;
-            }
-
-            console.log(`[Rockstar] Scanning game folder: ${entry}`);
-
-            // Look for .exe files in the game folder (deep scan)
-            const exeFiles = this.findExecutables(gamePath, 0, 20);
-            console.log(`[Rockstar] Found ${exeFiles.length} executables in ${entry}`);
-
-            // Filter out helper executables and launchers
-            const gameExes = exeFiles.filter(exe => {
-              const fileName = exe.toLowerCase();
-              const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-              return !fileNameOnly.includes('launcher') &&
-                !fileNameOnly.includes('uninstall') &&
-                !fileNameOnly.includes('unins') &&
-                !fileNameOnly.includes('setup') &&
-                !fileNameOnly.includes('installer') &&
-                !fileNameOnly.includes('socialclub') &&
-                !fileNameOnly.includes('social club') &&
-                !fileNameOnly.includes('redistributables') &&
-                !fileNameOnly.includes('redist') &&
-                !fileNameOnly.includes('updater') &&
-                !fileNameOnly.includes('gamelaunchhelper') &&
-                !fileNameOnly.includes('bootstrapper') &&
-                fileNameOnly !== 'crashreportclient.exe' &&
-                fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-                fileNameOnly !== 'battlenet.exe' &&
-                fileNameOnly !== 'battle.net.exe' &&
-                fileNameOnly !== 'crashpad_handler.exe' &&
-                fileNameOnly !== 'embark-crash-helper.exe' &&
-                fileNameOnly !== 'blizzardbrowser.exe' &&
-                fileNameOnly !== 'blizzarderror.exe';
-            });
-
-            console.log(`[Rockstar] Filtered to ${gameExes.length} game executables in ${entry}`);
-
-            if (gameExes.length > 0) {
-              // Prefer executables with the same name as the game folder
-              let mainExe = gameExes[0];
-
-              // Try to find an exe with the same name as the directory
-              const matchingExe = gameExes.find(exe => {
-                const exeName = exe.split(sep).pop()?.toLowerCase().replace('.exe', '') || '';
-                const gameName = entry.toLowerCase();
-                return exeName === gameName ||
-                  exeName.includes(gameName.replace(/\s+/g, '')) ||
-                  gameName.includes(exeName);
-              });
-
-              if (matchingExe) {
-                mainExe = matchingExe;
-              } else {
-                // Prefer executables closer to root (fewer directory separators)
-                const exeWithDepth = gameExes.map(exe => ({
-                  exe,
-                  depth: (exe.match(/[/\\]/g) || []).length
-                }));
-                exeWithDepth.sort((a, b) => a.depth - b.depth);
-                mainExe = exeWithDepth[0].exe;
-              }
-
-              console.log(`[Rockstar] ✓ Found Rockstar game: ${entry} (${mainExe})`);
-
-              results.push({
-                uuid: `rockstar-${entry}-${Date.now()}`,
-                source: 'rockstar' as const,
-                originalName: entry,
-                installPath: gamePath,
-                exePath: mainExe,
-                appId: undefined,
-                title: entry,
-                status: 'ambiguous' as const, // Rockstar games need metadata matching
-                isDownloading: this.gameFilteringService.isLikelyDownloading(gamePath),
-              });
-            } else {
-              console.log(`[Rockstar] No valid game executables found in ${entry}`);
-            }
-          } catch (err) {
-            console.warn(`[Rockstar] Could not scan folder "${entry}":`, err);
-            continue;
-          }
-        }
-      } catch (err) {
-        console.error('[Rockstar] Error scanning Rockstar Games directory:', err);
-      }
-
-      console.log(`[Rockstar] Total Rockstar games found: ${results.length}`);
-      return results;
-    } catch (error) {
-      console.error('[Rockstar] Error scanning Rockstar Games:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan a folder for manual executables
-   * Returns ScannedGameResult[] format for use with the importer
-   * Recursively scans all subdirectories and creates a game entry for each folder containing a valid executable
-   */
-  async scanFolderForExecutables(folderPath: string): Promise<ScannedGameResult[]> {
-    const results: ScannedGameResult[] = [];
-
-    try {
-      if (!existsSync(folderPath)) {
-        console.error(`[ImportService] Folder does not exist: ${folderPath}`);
-        return [];
-      }
-
-      console.log(`[ImportService] Scanning folder "${folderPath}" recursively for games...`);
-
-      // Find all executables recursively in the folder
-      const allExeFiles = this.findExecutables(folderPath, 0, 20);
-      console.log(`[ImportService] Found ${allExeFiles.length} total executables in "${folderPath}"`);
-
-      // Filter out helper executables
-      const gameExes = allExeFiles.filter(exe => {
-        const fileName = exe.toLowerCase();
-        const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-        return !fileNameOnly.includes('gamelaunchhelper') &&
-          !fileNameOnly.includes('bootstrapper') &&
-          !fileNameOnly.includes('uninstall') &&
-          !fileNameOnly.includes('unins') &&
-          !fileNameOnly.includes('setup') &&
-          !fileNameOnly.includes('installer') &&
-          !fileNameOnly.includes('install') &&
-          !fileNameOnly.includes('cleanup') &&
-          !fileNameOnly.includes('crashhandler') &&
-          !fileNameOnly.includes('redist') &&
-          !fileNameOnly.includes('directx') &&
-          !fileNameOnly.includes('updater') &&
-          !fileNameOnly.includes('launcher') &&
-          fileNameOnly !== 'crashreportclient.exe' &&
-          fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-          fileNameOnly !== 'battlenet.exe' &&
-          fileNameOnly !== 'battle.net.exe' &&
-          fileNameOnly !== 'cleanup.exe' &&
-          fileNameOnly !== 'crashpad_handler.exe' &&
-          fileNameOnly !== 'gamesessionmonitor.exe';
-      });
-
-      console.log(`[ImportService] Filtered to ${gameExes.length} valid game executables`);
-
-      if (gameExes.length === 0) {
-        console.log(`[ImportService] No valid game executables found in "${folderPath}"`);
-        return results;
-      }
-
-      // Group executables by their containing directory
-      // This way, if a folder has multiple executables, we create one game entry
-      // But if there are multiple folders each with their own executable, we create multiple games
-      const gamesByFolder = new Map<string, string[]>();
-
-      // Normalize the base folder path for comparison (use forward slashes consistently)
-      const normalizedBasePath = folderPath.replace(/\\/g, '/').toLowerCase();
-
-      for (const exePath of gameExes) {
-        const exeDir = dirname(exePath);
-        // Normalize the directory path (use forward slashes for consistency)
-        const normalizedDir = exeDir.replace(/\\/g, '/');
-
-        // Only include directories that are within the scanned folder
-        const normalizedExeDir = normalizedDir.toLowerCase();
-        if (!normalizedExeDir.startsWith(normalizedBasePath) && normalizedExeDir !== normalizedBasePath) {
-          console.log(`[ImportService] Skipping executable outside scanned folder: ${exePath} (dir: ${exeDir})`);
-          continue;
-        }
-
-        if (!gamesByFolder.has(normalizedDir)) {
-          gamesByFolder.set(normalizedDir, []);
-        }
-        gamesByFolder.get(normalizedDir)!.push(exePath);
-      }
-
-      console.log(`[ImportService] Found ${gamesByFolder.size} unique game folders`);
-
-      // Sort folders by depth (shallowest first) to detect parent/child relationships
-      const sortedFolders = Array.from(gamesByFolder.entries()).sort((a, b) => {
-        const depthA = a[0].split('/').length;
-        const depthB = b[0].split('/').length;
-        return depthA - depthB; // Shallowest first
-      });
-
-      // Track which folders we've added to avoid duplicates
-      const addedFolders = new Set<string>();
-
-      // Create a game entry for each folder
-      for (const [gameDir, exePaths] of sortedFolders) {
-        try {
-          // Check if this folder is a subfolder of an already added folder
-          // If so, skip it to avoid duplicates (e.g., Battlefield 6\SP when Battlefield 6 already exists)
-          const normalizedGameDir = gameDir.toLowerCase();
-          let isSubfolder = false;
-          for (const addedFolder of addedFolders) {
-            const normalizedAddedFolder = addedFolder.toLowerCase();
-            // Check if gameDir is a subfolder of addedFolder
-            if (normalizedGameDir.startsWith(normalizedAddedFolder + '/') ||
-              normalizedGameDir.startsWith(normalizedAddedFolder + '\\')) {
-              console.log(`[ImportService] Skipping subfolder game: ${gameDir} (parent folder already added: ${addedFolder})`);
-              isSubfolder = true;
-              break;
-            }
-          }
-
-          if (isSubfolder) {
-            continue;
-          }
-
-          // Use the folder name as the game title
-          const folderName = gameDir.split(/[/\\]/).pop() || 'Unknown';
-
-          console.log(`[ImportService] Processing game folder: ${folderName} (${gameDir}) with ${exePaths.length} executable(s)`);
-
-          // Select the best executable from this folder
-          let mainExe = exePaths[0];
-
-          // Prefer executables with the same name as the directory
-          const matchingExe = exePaths.find(exe => {
-            const exeName = exe.split(sep).pop()?.toLowerCase().replace('.exe', '') || '';
-            return exeName === folderName.toLowerCase();
-          });
-
-          if (matchingExe) {
-            mainExe = matchingExe;
-          } else {
-            // If no exact match, prefer executables closer to the folder root (fewer path separators)
-            const sortedByDepth = exePaths.sort((a, b) => {
-              // Count separators relative to the game directory
-              const depthA = a.substring(gameDir.length).split(sep).length;
-              const depthB = b.substring(gameDir.length).split(sep).length;
-              return depthA - depthB; // Prefer shallower (closer to folder root)
-            });
-            mainExe = sortedByDepth[0];
-            console.log(`[ImportService] Selected executable (closest to root): ${mainExe}`);
-          }
-
-          // Filter using shared game filtering service
-          const exeFileName = mainExe.split(/[/\\]/).pop() || '';
-          if (this.gameFilteringService.isLikelyNonGame({
-            name: folderName,
-            folderPath: gameDir,
-            exeName: exeFileName,
-            source: 'manual',
-          })) {
-            console.log(`[ImportService] Filtering out non-game from manual folder: ${folderName}`);
-            continue; // Skip this one
-          }
-
-          // Only create game if the directory is within or equal to the scanned folder
-          // This prevents creating games for parent directories
-          const normalizedFolderPath = folderPath.replace(/\\/g, '/');
-          const normalizedGameDirForCheck = gameDir.replace(/\\/g, '/');
-          if (normalizedGameDirForCheck.startsWith(normalizedFolderPath) || normalizedGameDirForCheck === normalizedFolderPath) {
-            const gameResult = {
-              uuid: `manual_folder-${gameDir}-${Date.now()}`,
-              source: 'Manual Folder',
-              originalName: folderName,
-              installPath: gameDir,
-              exePath: mainExe,
-              appId: undefined,
-              title: folderName,
-              status: 'ambiguous' as const, // Manual folder scans need metadata matching
-            };
-
-            // Special logging for John Wick Hex
-            if (folderName.toLowerCase().includes('john wick')) {
-              console.log(`[ImportService] ✓ Added John Wick Hex: installPath="${gameDir}", exePath="${mainExe}"`);
-            }
-
-            results.push(gameResult);
-            addedFolders.add(gameDir);
-            console.log(`[ImportService] ✓ Added game: ${folderName} from ${gameDir}`);
-          } else {
-            console.log(`[ImportService] Skipping game folder outside scanned path: ${gameDir} (scanned: ${folderPath})`);
-          }
-        } catch (err) {
-          console.warn(`[ImportService] Could not process game folder "${gameDir}":`, err);
-          continue;
-        }
-      }
-
-      console.log(`[ImportService] Scanned folder "${folderPath}" and found ${results.length} games`);
-      return results;
-    } catch (error) {
-      console.error(`[ImportService] Error scanning folder "${folderPath}":`, error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan EA App / Origin for installed games
-   * Games are typically in: {EAPath}\Games or subdirectories
-   */
-  private async scanEA(eaPath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[EA] Starting scan with path: ${eaPath}`);
-      const results: ScannedGameResult[] = [];
-
-      if (!existsSync(eaPath)) {
-        console.warn(`[EA] Path does not exist: ${eaPath}`);
-        return results;
-      }
-
-      // EA App games can be in multiple locations:
-      // 1. {EAPath}\Games (common)
-      // 2. Direct subdirectories of {EAPath}
-      // 3. {EAPath}\Program Files\EA Games (older Origin)
-
-      const possiblePaths = [
-        join(eaPath, 'Games'),
-        join(eaPath, 'Program Files', 'EA Games'),
-        join(eaPath, 'Program Files (x86)', 'EA Games'),
-        eaPath, // Scan the root path itself
-      ];
-
-      for (const gamesPath of possiblePaths) {
-        if (existsSync(gamesPath)) {
-          console.log(`[EA] Scanning: ${gamesPath}`);
-          const scanned = await this.scanGenericGamesFolder(gamesPath, 'ea');
-          results.push(...scanned);
-        }
-      }
-
-      console.log(`[EA] Found ${results.length} games total`);
-      return results;
-    } catch (error) {
-      console.error('[EA] Error scanning EA App:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan Battle.net for installed games
-   * Games are typically in: {BattlePath}\Games or subdirectories
-   */
-  private async scanBattle(battlePath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[Battle.net] Starting scan with path: ${battlePath}`);
-      const results: ScannedGameResult[] = [];
-
-      if (!battlePath || !existsSync(battlePath)) {
-        console.warn(`[Battle.net] Configured path does not exist or is empty: "${battlePath}". Will still attempt registry scan.`);
-      } else {
-        // 1. Check if the provided path is actually a game folder itself (common Blizzard structure)
-        if (existsSync(join(battlePath, '.build.info'))) {
-          console.log(`[Battle.net] Provided path looks like a Blizzard game folder: ${battlePath}`);
-          const scanned = await this.scanGenericGamesFolder(battlePath, 'battle', false);
-          results.push(...scanned);
-        } else {
-          // 2. Battle.net games are typically ONLY in the Games subdirectory
-          // Don't scan the root path to avoid picking up the launcher itself
-          const gamesPath = join(battlePath, 'Games');
-
-          if (existsSync(gamesPath)) {
-            console.log(`[Battle.net] Scanning games directory: ${gamesPath}`);
-            const scanned = await this.scanGenericGamesFolder(gamesPath, 'battle');
-            results.push(...scanned);
-          } else {
-            console.log(`[Battle.net] Games directory not found at: ${gamesPath}`);
-          }
-
-          // 3. Improved check: If the path provided is the Battle.net launcher folder, check the parent
-          // Many users point to "C:\\Program Files (x86)\\Battle.net" but games are in "C:\\Program Files (x86)"
-          const isLauncherDir = existsSync(join(battlePath, 'Battle.net.exe')) ||
-            battlePath.toLowerCase().endsWith('battle.net') ||
-            battlePath.toLowerCase().endsWith('battlenet');
-
-          if (isLauncherDir) {
-            const parentDir = dirname(battlePath);
-            if (existsSync(parentDir) && parentDir !== battlePath) {
-              console.log(`[Battle.net] Detected launcher directory, also scanning parent: ${parentDir}`);
-              try {
-                const parentEntries = readdirSync(parentDir, { withFileTypes: true });
-                for (const entry of parentEntries) {
-                  if (entry.isDirectory()) {
-                    const dirName = entry.name.toLowerCase();
-                    // Skip the launcher itself and other system folders
-                    if (dirName === 'battle.net' || dirName === 'battlenet' || dirName === 'blizzard' || dirName === 'common') continue;
-
-                    const potentialGamePath = join(parentDir, entry.name);
-                    // Check if it's a Blizzard game (has .build.info)
-                    if (existsSync(join(potentialGamePath, '.build.info'))) {
-                      // Avoid duplicate if already found
-                      if (results.some(r => r.installPath === potentialGamePath)) continue;
-
-                      console.log(`[Battle.net] Found Blizzard game via .build.info in parent: ${entry.name}`);
-                      const scanned = await this.scanGenericGamesFolder(potentialGamePath, 'battle', false);
-                      results.push(...scanned);
-                    }
-                  }
-                }
-              } catch (err) {
-                console.warn(`[Battle.net] Could not scan parent directory ${parentDir}:`, err);
-              }
-            }
-          }
-
-          // 4. Additional check for subdirectories (existing logic)
-          const entries = readdirSync(battlePath, { withFileTypes: true });
-          for (const entry of entries) {
-            if (entry.isDirectory()) {
-              const dirName = entry.name.toLowerCase();
-
-              // Skip already handled and irrelevant folders
-              if (dirName === 'battle.net' ||
-                dirName === 'battlenet' ||
-                dirName === 'launcher' ||
-                dirName === 'blizzard' ||
-                dirName === 'system' ||
-                dirName === 'cache' ||
-                dirName === 'logs' ||
-                dirName === 'temp' ||
-                dirName === 'games') {
-                continue;
-              }
-
-              const potentialGamePath = join(battlePath, entry.name);
-              // Avoid duplicate
-              if (results.some(r => r.installPath === potentialGamePath)) continue;
-
-              const scanned = await this.scanGenericGamesFolder(potentialGamePath, 'battle', false);
-              results.push(...scanned);
-            }
-          }
-        } // End of inner else
-      } // End of outer else
-
-      // 5. Query Windows Registry for ANY Blizzard games installed anywhere
-      if (typeof process !== 'undefined' && process.platform === 'win32') {
-        try {
-          console.log(`[Battle.net] Checking registry for installed games...`);
-          const keysToSearch = [
-            'HKLM:\\SOFTWARE\\WOW6432Node\\Microsoft\\Windows\\CurrentVersion\\Uninstall',
-            'HKLM:\\SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Uninstall'
-          ];
-
-          for (const baseKey of keysToSearch) {
-            try {
-              // Get all keys where Publisher is Blizzard Entertainment or Battle.net
-              // Use -EncodedCommand to avoid $_ being consumed by cmd.exe shell
-              const psScript = `Get-ItemProperty '${baseKey}\\*' -ErrorAction SilentlyContinue | Where-Object { $_.Publisher -match 'Blizzard' -or $_.Publisher -match 'Battle.net' -or $_.UninstallString -match 'Blizzard Uninstaller' } | Select-Object DisplayName, InstallLocation | ConvertTo-Json -Compress`;
-              const encodedCmd = Buffer.from(psScript, 'utf16le').toString('base64');
-              const output = execSync(`powershell -NoProfile -NonInteractive -EncodedCommand ${encodedCmd}`, { encoding: 'utf-8', stdio: ['pipe', 'pipe', 'ignore'] }).trim();
-
-              if (output) {
-                const gamesArr = output.startsWith('[') ? JSON.parse(output) : [JSON.parse(output)];
-                for (const game of gamesArr) {
-                  if (game.DisplayName && game.InstallLocation) {
-                    const name = game.DisplayName;
-                    const path = game.InstallLocation;
-
-                    // Skip the launcher itself
-                    if (name === 'Battle.net' || path.toLowerCase().includes('battle.net')) continue;
-
-                    if (!results.some(r => r.installPath === path) && existsSync(path)) {
-                      console.log(`[Battle.net] Found game via registry: ${name} at ${path}`);
-                      const scanned = await this.scanGenericGamesFolder(path, 'battle', false);
-                      results.push(...scanned);
-                    }
-                  }
-                }
-              }
-            } catch (e) {
-              // Ignore if registry read fails
-            }
-          }
-        } catch (err) {
-          console.warn(`[Battle.net] Registry scan failed:`, err);
-        }
-      }
-
-      console.log(`[Battle.net] Found ${results.length} games total`);
-      return results;
-    } catch (error) {
-      console.error('[Battle.net] Error scanning Battle.net:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan Humble App for installed games
-   * Games are typically in subdirectories
-   */
-  private async scanHumble(humblePath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[Humble] Starting scan with path: ${humblePath}`);
-      const results: ScannedGameResult[] = [];
-
-      if (!existsSync(humblePath)) {
-        console.warn(`[Humble] Path does not exist: ${humblePath}`);
-        return results;
-      }
-
-      // Humble games are typically in subdirectories
-      console.log(`[Humble] Scanning: ${humblePath}`);
-      const scanned = await this.scanGenericGamesFolder(humblePath, 'humble');
-      results.push(...scanned);
-
-      console.log(`[Humble] Found ${results.length} games total`);
-      return results;
-    } catch (error) {
-      console.error('[Humble] Error scanning Humble:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Scan itch.io for installed games
-   * Games are typically in subdirectories
-   */
-  private async scanItch(itchPath: string): Promise<ScannedGameResult[]> {
-    try {
-      console.log(`[itch.io] Starting scan with path: ${itchPath}`);
-      const results: ScannedGameResult[] = [];
-
-      if (!existsSync(itchPath)) {
-        console.warn(`[itch.io] Path does not exist: ${itchPath}`);
-        return results;
-      }
-
-      // itch.io games are typically in subdirectories
-      console.log(`[itch.io] Scanning: ${itchPath}`);
-      const scanned = await this.scanGenericGamesFolder(itchPath, 'itch');
-      results.push(...scanned);
-
-      console.log(`[itch.io] Found ${results.length} games total`);
-      return results;
-    } catch (error) {
-      console.error('[itch.io] Error scanning itch.io:', error);
-      return [];
-    }
-  }
-
-  /**
-   * Generic method to scan a games folder for any launcher
-   * Recursively scans all subdirectories and creates a game entry for each folder containing a valid executable
-   */
-  private async scanGenericGamesFolder(gamesPath: string, source: 'ea' | 'battle' | 'humble' | 'itch', isLibraryRoot: boolean = true): Promise<ScannedGameResult[]> {
-    const results: ScannedGameResult[] = [];
-
-    try {
-      console.log(`[${source}] Scanning games folder: ${gamesPath}`);
-
-      // Find all executables recursively in the folder
-      const allExeFiles = this.findExecutables(gamesPath, 0, 20);
-      console.log(`[${source}] Found ${allExeFiles.length} total executables`);
-
-      // Filter out helper executables
-      const gameExes = allExeFiles.filter(exe => {
-        const fileName = exe.toLowerCase();
-        const fileNameOnly = fileName.split(/[/\\]/).pop() || '';
-        return !fileNameOnly.includes('gamelaunchhelper') &&
-          !fileNameOnly.includes('bootstrapper') &&
-          !fileNameOnly.includes('uninstall') &&
-          !fileNameOnly.includes('unins') &&
-          !fileNameOnly.includes('setup') &&
-          !fileNameOnly.includes('installer') &&
-          !fileNameOnly.includes('launcher') &&
-          !fileNameOnly.includes('updater') &&
-          !fileNameOnly.includes('redist') &&
-          !fileNameOnly.includes('directx') &&
-          fileNameOnly !== 'crashreportclient.exe' &&
-          fileNameOnly !== 'battlenet.overlay.runtime.exe' &&
-          fileNameOnly !== 'battlenet.exe' &&
-          fileNameOnly !== 'battle.net.exe' &&
-          fileNameOnly !== 'cleanup.exe' &&
-          fileNameOnly !== 'crashpad_handler.exe' &&
-          fileNameOnly !== 'gamesessionmonitor.exe';
-      });
-
-      console.log(`[${source}] Filtered to ${gameExes.length} valid game executables`);
-
-      if (gameExes.length === 0) {
-        return results;
-      }
-
-      // Group executables by their containing directory
-      const gamesByFolder = new Map<string, string[]>();
-
-      for (const exePath of gameExes) {
-        let gameDir: string;
-
-        if (isLibraryRoot) {
-          const relativePath = exePath.substring(gamesPath.length);
-          const parts = relativePath.split(/[/\\]/).filter(Boolean);
-          if (parts.length > 0) {
-            gameDir = join(gamesPath, parts[0]);
-          } else {
-            gameDir = gamesPath;
-          }
-        } else {
-          gameDir = gamesPath;
-        }
-
-        const normalizedDir = gameDir.replace(/\\/g, sep).replace(/\/\//g, '/');
-
-        // Skip known non-game directories
-        const dirName = normalizedDir.split(sep).pop()?.toLowerCase() || '';
-        if (dirName === 'launcher' ||
-          dirName === 'redistributables' ||
-          dirName === 'support' ||
-          dirName === 'tools' ||
-          dirName.startsWith('$')) {
-          continue;
-        }
-
-        if (!gamesByFolder.has(normalizedDir)) {
-          gamesByFolder.set(normalizedDir, []);
-        }
-        gamesByFolder.get(normalizedDir)!.push(exePath);
-      }
-
-      console.log(`[${source}] Found ${gamesByFolder.size} unique game folders`);
-
-      // Sort folders by depth (shallowest first) to detect parent/child relationships
-      const sortedFolders = Array.from(gamesByFolder.entries()).sort((a, b) => {
-        const depthA = a[0].split('/').length;
-        const depthB = b[0].split('/').length;
-        return depthA - depthB; // Shallowest first
-      });
-
-      // Track which folders we've added to avoid duplicates
-      const addedFolders = new Set<string>();
-
-      // Create a game entry for each folder
-      for (const [gameDir, exePaths] of sortedFolders) {
-        try {
-          // Only process directories within the games path
-          const normalizedGamesPath = gamesPath.replace(/\\/g, '/');
-          const normalizedGameDir = gameDir.replace(/\\/g, '/');
-          if (!normalizedGameDir.startsWith(normalizedGamesPath) && normalizedGameDir !== normalizedGamesPath) {
-            continue;
-          }
-
-          // Check if this folder is a subfolder of an already added folder
-          // If so, skip it to avoid duplicates (e.g., Battlefield 6\SP when Battlefield 6 already exists)
-          const normalizedGameDirLower = normalizedGameDir.toLowerCase();
-          let isSubfolder = false;
-          for (const addedFolder of addedFolders) {
-            const normalizedAddedFolder = addedFolder.toLowerCase();
-            // Check if gameDir is a subfolder of addedFolder
-            if (normalizedGameDirLower.startsWith(normalizedAddedFolder + '/') ||
-              normalizedGameDirLower.startsWith(normalizedAddedFolder + '\\')) {
-              console.log(`[${source}] Skipping subfolder game: ${gameDir} (parent folder already added: ${addedFolder})`);
-              isSubfolder = true;
-              break;
-            }
-          }
-
-          if (isSubfolder) {
-            continue;
-          }
-
-          // Use the folder name as the game title
-          const folderName = gameDir.split(/[/\\]/).pop() || 'Unknown';
-
-          // Select the best executable from this folder
-          let mainExe = exePaths[0];
-
-          // Prefer executables with the same name as the directory
-          const matchingExe = exePaths.find(exe => {
-            const exeName = exe.split(sep).pop()?.toLowerCase().replace('.exe', '') || '';
-            return exeName === folderName.toLowerCase();
-          });
-
-          if (matchingExe) {
-            mainExe = matchingExe;
-          } else {
-            // Prefer executables closer to the folder root
-            const sortedByDepth = exePaths.sort((a, b) => {
-              const depthA = a.substring(gameDir.length).split(sep).length;
-              const depthB = b.substring(gameDir.length).split(sep).length;
-              return depthA - depthB;
-            });
-            mainExe = sortedByDepth[0];
-          }
-
-          // Filter using shared game filtering service
-          const exeFileName = mainExe.split(/[/\\]/).pop() || '';
-          if (this.gameFilteringService.isLikelyNonGame({
-            name: folderName,
-            folderPath: gameDir,
-            exeName: exeFileName,
-            source: source as any,
-          })) {
-            console.log(`[${source}] Filtering out non-game: ${folderName}`);
-            continue; // Skip this one
-          }
-
-          results.push({
-            uuid: `${source}-${gameDir}-${Date.now()}`,
-            source: source,
-            originalName: folderName,
-            installPath: gameDir,
-            exePath: mainExe,
-            appId: undefined,
-            title: folderName,
-            status: 'ambiguous' as const,
-            isDownloading: this.gameFilteringService.isLikelyDownloading(gameDir),
-          });
-          addedFolders.add(normalizedGameDir);
-          console.log(`[${source}] ✓ Added game: ${folderName} from ${gameDir}${this.gameFilteringService.isLikelyDownloading(gameDir) ? ' (Downloading)' : ''}`);
-        } catch (err) {
-          console.warn(`[${source}] Could not process game folder "${gameDir}":`, err);
-          continue;
-        }
-      }
-
-      console.log(`[${source}] Found ${results.length} games total`);
-    } catch (err) {
-      console.error(`[${source}] Error scanning games folder:`, err);
     }
 
     return results;
