+++ main/scanners/GameScanner.ts
@@ -1,0 +1,7 @@
+import { ScannedGameResult } from './ScannedGameResult.js';
+
+export { ScannedGameResult };
+
+export interface GameScanner {
+  scan(path: string): Promise<ScannedGameResult[]>;
+}
